Allow Drop Frame Timecode in DM XML "durations"
This also normalises this type with that used for timecodes, in that it
now allows a maximum of 5 digits for the frames part. Previously it
allowed any number of digits for that part.

It also fixes a small bug in the regular expression pattern for durations
that allowed the field to be empty.

Resolves AMWA-TV/AS-11_Overview#45